first steps

+++ subscriptions/models.py
@@ -41,11 +41,12 @@ class Subscription(models.Model):
     name = models.CharField(
         'Название', max_length=LENGTH_LIMITS_CHAR_FIELDS)
     description = models.TextField('Описание')
-    monthly_price = models.FloatField(
+    monthly_price = models.DecimalField(
         'Цена подписки за месяц', decimal_places=2)
-    semi_annual_price = models.FloatField(
+    semi_annual_price = models.DecimalField(
         'Цена подписки за полгода', decimal_places=2)
-    annual_price = models.FloatField('Цена подписки за год', decimal_places=2)
+    annual_price = models.DecimalField(
+        'Цена подписки за год', decimal_places=2)
     users = models.ManyToManyField(
         User,
         through='UserSubscription',
